Mypy cover fixes

+++ src/boneio/helper/state_manager.py
@@ -16,8 +16,8 @@ _LOGGER = logging.getLogger(__name__)
 
 
 class CoverStateEntry(BaseModel):
-    position: float
-    tilt: float | None = None
+    position: int
+    tilt: int | None = None
 
 
 class State(BaseModel):
